Tidied Portal Logging
